Subject add css Problem (72 characters max) Solution Note
[Ticket: X]
